move settings file and make some changes for signature
