+++ employeeRecognition/ClientApp/src/components/Users/AddUser.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Redirect } from 'react-router-dom';
-//import axios from 'axios';
+import axios from 'axios';
 var util = require('util');
 
 export class AddUser extends Component {
@@ -145,7 +145,6 @@ export class AddUser extends Component {
                                 name="first_name"
                                 placeholder="First Name"
                                 autoFocus
-                                required
                             />
                         </div>
                         <div className="form-group">
@@ -157,7 +156,6 @@ export class AddUser extends Component {
                                 onChange={this.onChange}
                                 name="last_name"
                                 placeholder="Last Name"
-                                required
                             />
                         </div>
                         <div className="form-group">
@@ -169,7 +167,6 @@ export class AddUser extends Component {
                                 onChange={this.onChange}
                                 name="email"
                                 placeholder="Email"
-                                required
                             />
                         </div>
                         <div className="form-group">
@@ -181,7 +178,6 @@ export class AddUser extends Component {
                                 onChange={this.onChange}
                                 name="password"
                                 placeholder="Password"
-                                required
                             />
                         </div>
                         <div className="form-group">
